+++ src/test/java/com/meetup/server/event/application/EventServiceTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -39,7 +40,7 @@ class EventServiceTest {
     void 비로그인_사용자가_이벤트를_생성한다() {
         StartPointRequest startPointRequest = new StartPointRequest(
                 null, "선정릉역 수인분당선", "서울특별시 강남구 삼성동 111-114", "서울특별시 강남구 선릉로 지하580",
-                127.043999, 37.510297
+                BigDecimal.valueOf(127.043999), BigDecimal.valueOf(37.510297)
         );
         UUID eventId = eventService.createEvent(null, startPointRequest);
 
@@ -50,7 +51,7 @@ class EventServiceTest {
     void 로그인_사용자가_이벤트를_생성한다() {
         StartPointRequest startPointRequest = new StartPointRequest(
                 "땡수팟", "선정릉역 수인분당선", "서울특별시 강남구 삼성동 111-114", "서울특별시 강남구 선릉로 지하580",
-                127.043999, 37.510297
+                BigDecimal.valueOf(127.043999), BigDecimal.valueOf(37.510297)
         );
         UUID eventId = eventService.createEvent(user.getUserId(), startPointRequest);
